requirements.txt
click
pandas
python-dotenv
openai>=1.40.0
